requirements.txt
asgiref==3.6.0
Django==4.1.7
django-admin-interface==0.24.2
django-colorfield==0.8.0
django-jalali-date==1.0.1
jdatetime==4.1.0
Pillow==9.4.0
psycopg2-binary==2.9.5
sqlparse==0.4.3
